Creating cases.rs and moving functions around
